pyproject.toml
[project]
name = "ai-sportsbettor"
version = "0.1.0"
description = "Sports news ingestion and odds edge detection pipeline"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "requests>=2.32.0",
    "feedparser>=6.0.11",
    "beautifulsoup4>=4.12.3",
    "python-dateutil>=2.9.0",
    "pandas>=2.2.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "sqlalchemy>=2.0.30",
    "psycopg[binary]>=3.1.19",
    "alembic>=1.13.1",
    "apscheduler>=3.10.4",
    "python-dotenv>=1.0.1",
    "google-cloud-storage>=2.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "ruff>=0.4.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

